move internal.OverlayStack to app.OverlayStack
… to avoid dependencies down the package tree.
